Volatile qualifier added for bare metal portevent.c static variables (#50)
Volatile qualifier is required for variables accessed by functions which are called from an interrupts. Code become broken at higher compiler optimization level if volatile qualifier is not set.

While integrating bare metal code, with -O3 -flto GCC optimization settings, both AVR and ARM architectures refused to receive Modbus RTU requests. Debugging discovered that eQueuedEvent and xEventInQueue were optimized and was not tested upon an event.

This behavior of original code is correct for ANSI C - the compiler can optimize the variables which are not changed in current context. But in portevent.c both variables are accessed from a function which is called from an interrupt - which is out of compiler context. So in portevent.c both variables, eQueuedEvent and xEventInQueue, should be defined as volatile. This is exactly the case for which the qualifier is designed.